bit-dropping census: a fail-closed repair turned a silent double-spend into a free consensus halt
The operator, after the fifth instance in a day: *"we need to stop dropping bits."* This is the census
and the top fixes. **The worst finding is not one of the fixes.**

## ⚑ TOP OPEN ITEM — `cells_root`: a $0-cost permanent liveness kill on the consensus anchor

`heap_addr(CELLS_COLLECTION, hash_bytes(id))` is **one felt**. The 2026-07-28 repair made
`assert_addr_unique` **panic, release-active**. And `CellId::derive_raw` is BLAKE3 over attacker-chosen
`(public_key, token_id)` — **grindable entirely OFFLINE.**

**Grind ~2^15.5 (≈46,000 hashes, milliseconds), create TWO cells, and `cells_root` panics on every turn
forever.** Targeted at a specific victim cell it is 2^31 and one creation.

⚑ **The fail-closed repair converted a silent double-spend into a free consensus halt**, and the key
widening was deferred to the MapOp epoch. Fail-closed is right; fail-closed *on an attacker-chosen
collidable key* is a denial-of-service primitive.

## Fixed here

**`ShieldedNoteSet` — the last accumulator sibling on the 31-bit linear fold.** Its leaf was
`fold_bytes32_to_bb(commitment)` with the value column **ZERO by design**, so that ~31-bit address was
the leaf's *entire content*: **two distinct shielded notes were literally the same leaf.** Its three
siblings migrated on 07-31; this one was left behind and was the worst of the four. Now
`exact_linked_append_root8` under `FSI2`/`FSN2`/`FSE2` — 16 injective u16 limbs, 2^256, no reduction,
hiding preserved. Landed with a **falsifier** that constructs the collision at O(1) (chunk 0 = `0` vs
`p`), asserts it genuinely collides under the retired fold, and asserts the new root separates it.

**`compute_revealed_terms_commitment` — 30 bits per disclosed term.** Each 32-byte term went
`blake3 → 4 bytes → & ((1<<30)−1)`, and that felt was its only binding, so a **2^15 birthday grind over
attacker-chosen attribute bytes bought a substituted cleartext disclosure.** ⚠ The 8-felt `WideHash`
above it was irrelevant — **the waist was upstream of the sponge.** Now absorbs `field_limbs9`; nine and
not eight is forced by pigeonhole. Domain bumped to `-v2` so stale commitments refuse.

**Two wasm32 divergences.** `hash_index` fed `usize::to_le_bytes()` into a BLAKE3 preimage — **4 bytes
on wasm32, 8 on x86_64**, so a wasm light client and a native prover built *provably different Merkle
trees* from identical inputs, with nothing red. `shard_for` discarded 4 of 8 digest bytes on wasm32
before its modulus, changing `combined_root`.

**And the gate's alias hole**: `fold_bytes32` / `fold_value32` are **byte-identical bodies under
different names** and were invisible to the rule for its entire life. Added — and it immediately reds
`cell/src/commitment.rs:561`, the cap-leaf target **in the deployed canonical state commitment**, in a
file the gate already scanned.

## The rest of the (A) list, ranked

**`iroot` — the receipt-log MMR is ONE FELT inside the "faithful 8-felt" anchor.** Two receipt logs
colliding at ~2^15.5 give a byte-identical signed `pre/post_state_hash`. ⚠ And
`rotation_witness.rs:349` still cites `mroot_injective` as making the root "bind the WHOLE log" —
**that theorem was DELETED 2026-07-28 as pigeonhole-false.** The honest form is `mroot_binds_or_collides`
at ≈2^15.5.

**The key octet: `A` and `−A` pack identically — collision cost ZERO.** `canonical_32_to_felts_8` drops
bits 6–7 of bytes 3,7,…,31, and the Ed25519 x-sign is among them. ⚑ **`KeyLanes9` is already authored
and proved in Lean, with a machine-checked left inverse — and not emitted**, its own doc saying "Not
this lane's edit." Cost: `rotatedNumPreLimbs` 184→187, re-emit, VK, epoch 16→17. **Four byte-identical
Rust copies, no cross-check test.**

**`hash_bytes` is non-injective on input LENGTH, at zero cost** — verified empirically:
`hash_bytes(b"foo") == hash_bytes(b"foo\0")`, because the domain tag is the *felt* count, not the byte
count. Bites `sandstorm-bridge::var_addr` (serve `value ‖ "\0"` and inclusion passes).

## ⚠ And the gate verdict, which is uncomfortable

**The degraded-felt gate is not a gate today: it is RED with 12 errors, so nothing new can turn it red.**
It matches a **name, not a shape**; `[X; 8]` misses mutate-after-zero-init (`wide_from_felt` writes
`wide[0] = felt` into a zeroed array, in a file it already scans); and `byte-to-felt-arithmetic` is
`severity: warning` and therefore *cannot fail*.

Honest verdict: **width is gate-able by a type wall; the two things that actually cost this tree — a
non-injective encoder, and a bare map/sort KEY — are not gate-able syntactically.** Every kind-D wound
is a key, and no type has ever guarded one.

## What broke

`ShieldedNoteSet::root8` values change and the root is now **order-dependent by construction**, so an
order-independence assertion that was **false** is rewritten with a non-vacuity leg proving the
dependence. Credential presentations re-emit at `-v2`. `hash_index`/`shard_for` change **on wasm32
only** — they were already wrong there. No VK rotates. Persisted shielded roots re-genesis.

+++ sdk/src/cipherclerk.rs
@@ -4958,10 +4958,28 @@ impl AgentCipherclerk {
     }
 
     /// Derive a deterministic sibling hash for Merkle path construction.
+    ///
+    /// ⚑ **`level` and `sibling_idx` are absorbed as `u64`, NOT as `usize`** (fixed
+    /// 2026-08-01). `usize::to_le_bytes()` emits **4 bytes on `wasm32` and 8 on `x86_64`**, so
+    /// the previous body gave the BLAKE3 sponge a different preimage per target for the same
+    /// `(level, sibling_idx, key)` — and therefore a different sibling hash, a different
+    /// Merkle path and a different recomputed root. A wasm light client and a native prover
+    /// built provably different trees over identical inputs, with nothing red on either.
+    /// "Deterministic" in the line above was true per-target and false across them.
+    ///
+    /// This is the discipline `circuit-prove/src/ivc_turn_chain.rs` already states for the
+    /// IVC count lane ("BEFORE any `as usize` (which is 32 bits under `wasm32`, where a
+    /// `usize`-typed guard would have been truncated past)"): a `usize` may index, but it may
+    /// never reach a hash preimage or a committed width unwrapped.
+    ///
+    /// ⚠ The `% BABYBEAR_P` on the low four digest bytes below is a SEPARATE, unfixed
+    /// narrowing — a ~31-bit image of a 256-bit digest, aliasing `x` with `x + p` on 53.1% of
+    /// draws. It is not repaired here because the value's consumer is the Poseidon2 path
+    /// recompute whose felt width is descriptor-fixed; widening it is the map-key epoch.
     fn hash_index(level: usize, sibling_idx: usize, key: &[u8; 32]) -> u32 {
         let mut hasher = blake3::Hasher::new();
-        hasher.update(&level.to_le_bytes());
-        hasher.update(&sibling_idx.to_le_bytes());
+        hasher.update(&(level as u64).to_le_bytes());
+        hasher.update(&(sibling_idx as u64).to_le_bytes());
         hasher.update(key);
         let hash = hasher.finalize();
         let bytes = hash.as_bytes();
